PathFinder now paths in running-steps instead of walking-steps

+++ src/main/java/wiki/runescape/oldschool/pathfinder/logic/GraphEdge.java
@@ -1,5 +1,5 @@
 package wiki.runescape.oldschool.pathfinder.logic;
 
-public record GraphEdge(GraphVertex to, byte cost, String methodOfMovement) {
+public record GraphEdge(GraphVertex to, float cost, String methodOfMovement) {
 
 }
